Update lang.js::_.getPluralForm
In our system, and I'm sure many others. We use a suffix country code when setting our locale for certain regions. 
ie. en_US, en_CA, en_GB.

The switch statement is based on the language code and not the locale. It should take the first two characters of the locale instead (the language code.)